+++ src/server/physics/event/ig_ENTITY_OneTimeSwitch.ts
@@ -1,0 +1,20 @@
+import { prestart } from '../../../loading-stages'
+import { isPhysics } from '../is-physics-server'
+import { setNextSetBy, unsetNextSetBy } from './vars'
+
+prestart(() => {
+    if (!PHYSICS) return
+
+    ig.ENTITY.OneTimeSwitch.inject({
+        ballHit(ball) {
+            if (!isPhysics(multi.server)) return this.parent(ball)
+
+            if ('combatant' in ball && ball.combatant && ball.combatant instanceof ig.Entity)
+                setNextSetBy(ball.getCombatantRoot())
+
+            const ret = this.parent(ball)
+            unsetNextSetBy()
+            return ret
+        },
+    })
+})