+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add ping interval and ping timeout server options
 - Add remote camera and screen blur step synchronization
+- Add ig.ENTITY.OneTimeSwitch event chain preservation
 
 ### Fixed